Backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
sqlalchemy
alembic          # optional for DB migrations
pydantic
